chore: split out getShouldRenderIntegrationsTab
